refactor(deployments): rename app-card.component files to deployment-card.component
related to openshiftio/openshift.io#1382
